% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeDsWs.R
\name{computeDsWs}
\alias{computeDsWs}
\title{A function that computes D_ijtk and W_ijtk using the held-out trade data}
\usage{
computeDsWs(file.dir)
}
\arguments{
\item{file.dir}{The file directory of the held-out data.}
}
\value{
A dataframe that contains D_ijtk and W_ijtk in long form.
}
\description{
A function that computes D_ijtk and W_ijtk using the held-out trade data
}
\examples{
file.dir <- "../dynCluster-master/example/toy/dynamic_1970.csv"
df <- computeDsWs(file.dir)
}
