% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLogLik.R
\name{computeLogLik}
\alias{computeLogLik}
\title{A function to compute hold-out log-likelihood}
\usage{
computeLogLik(n.cluster, n.period, file.dir, test.df, test.run)
}
\arguments{
\item{n.cluster}{The number of clusters in the fitted model (integer).}

\item{n.period}{The number of periods in the fitted model (integer).}

\item{file.dir}{The file directory of dynCluster outputs.}

\item{test.df}{A dataframe that contains the hold-out data (D_ijtk and W_ijtk) 
created using the function computeDsWs.}

\item{test.run}{A logical TRUE/FALSE statement indicating whether to conduct a quick test 
run based on only two dyads (100_104 and 156_840).}
}
\value{
A list object that contains a scalar for the overall hold-out log-likelihood
and a dataframe that contains the hold-out log-likelihood for each dyad.
}
\description{
A function to compute hold-out log-likelihood
}
\examples{
\dontrun{# set directories
test.dir <- "./example/toy/dynamic_1970.csv"
out.dir <- "./example/toy"

# extract D_ijtk and W_ijtk
df <- computeDsWs(test.dir)

# compute hold-out log-likelihood
system.time(
  out <- computeLogLik(n.cluster = 3, n.period = 10, 
                       file.dir = out.dir, test.df = df, 
                       test.run = TRUE)
)}
}
