% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztm.R
\name{mainZTM}
\alias{mainZTM}
\title{Title This is a function that calls the CPP ZTM code via a config.txt which specifies parameters and data}
\usage{
mainZTM(dataDirectory = "./", comeBack = FALSE)
}
\arguments{
\item{comeBack}{}
}
\value{
1 for error and 0 for success
}
\description{
Title This is a function that calls the CPP ZTM code via a config.txt which specifies parameters and data
}

