% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztm.R
\name{testExample}
\alias{testExample}
\title{Title This is an example of using ztmR, assuming the required data is in the dataDirectory}
\usage{
testExample(dataDirectory = "./", nThreads = 6, comeBack = FALSE)
}
\arguments{
\item{comeBack}{}
}
\value{
1 for error and 0 for success
}
\description{
Title This is an example of using ztmR, assuming the required data is in the dataDirectory
}

