% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztm.R
\name{mainZTM}
\alias{mainZTM}
\title{Title This is a function that calls the CPP ZTM code via a config.txt which specifies parameters and data}
\usage{
mainZTM(dataDirectory = "./", comeBack = FALSE)
}
\arguments{
\item{dataDirectory}{the working directory contains the data}

\item{comeBack}{return from the working directory}
}
\value{
1 for error and 0 for success
}
\description{
Title This is a function that calls the CPP ZTM code via a config.txt which specifies parameters and data
}
\examples{
library(Rcpp)
Sys.setenv("PKG_CXXFLAGS" = "-fopenmp -lboost_regex")
Sys.setenv("PKG_LIBS" = "-fopenmp -lboost_regex")
\dontrun{sourceCpp("./src/mainRcpp.cpp")}
\dontrun{mainZTM("./example", comeBack=TRUE)}
}

